fix: StoreProductRequest and UpdateProductRequest

+++ app/Http/Requests/Comment/StoreCommentRequest.php
@@ -24,7 +24,7 @@ class StoreCommentRequest extends FormRequest
         return [
             'message' => ['required', 'string'],
             'commentable_type' => ['required'],
-            'commentable_id' => ['required', 'int'],
+            'commentable_id' => ['required', 'numeric'],
         ];
     }
 }
